+++ .idea/deploymentTargetSelector.xml
@@ -4,11 +4,10 @@
     <selectionStates>
       <SelectionState runConfigName="app">
         <option name="selectionMode" value="DROPDOWN" />
-        <DropdownSelection timestamp="2025-06-03T00:12:14.196696700Z">
+        <DropdownSelection timestamp="2025-06-11T20:58:47.352162200Z">
           <Target type="DEFAULT_BOOT">
             <handle>
               <DeviceId pluginId="LocalEmulator" identifier="path=C:\Users\kevin\.android\avd\Medium_Phone.avd" />
-
             </handle>
           </Target>
         </DropdownSelection>
